split tests into another path
