+++ nightmare/jest.e2e.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   name: 'E2E',
   displayName: 'E2E',
-  testMatch: ['**/nightmare/**/paths/metadata.spec.js?(x)'],
+  testMatch: ['**/nightmare/**/paths/metadataprops.spec.js?(x)'],
   testEnvironment: 'node',
   setupFilesAfterEnv: ['<rootDir>/setUpJestE2E.js'],
 };
